feat: Gameplay mechanics implemented - XP level counter connected - Avatar switching according to levels - History is displayed based on task completion time, not end time

+++ src/main/java/org/example/controller/TaskController.java
@@ -1,5 +1,6 @@
 package org.example.controller;
 
+import org.example.Game.GameController;
 import org.example.model.Task;
 import org.example.model.TaskRepository;
 
@@ -8,16 +9,16 @@ import java.util.List;
 public class TaskController {
 
     private final TaskRepository repository;
+    private final GameController gameController;
 
-
-    public TaskController(TaskRepository repository) {
+    public TaskController(TaskRepository repository, GameController gameController) {
         this.repository = repository;
+        this.gameController = gameController;
     }
 
     public List<Task> getAllTasks(){
         return repository.findAll();
     }
-
 
     public void addTask(String title, String category){
         Task newTask = new Task(title, category);
@@ -29,13 +30,27 @@ public class TaskController {
     }
 
     public List<Task> getAllHistoryTasks(){
-        return repository.findAll();
+        return repository.findAllHistory();
     }
 
     public void deleteTask(int id){
-        repository.deleteById(id);
-    }
+        Task task = findTaskById(id);
 
+
+        if (task != null) {
+            String cat = task.getCategory();
+
+            switch (cat){
+                case "Красная зона" -> gameController.addXPToCharacter(40);
+                case "Зеленая зона" -> gameController.addXPToCharacter(30);
+                case "Синяя зона" -> gameController.addXPToCharacter(20);
+                case "Желтая зона" -> gameController.addXPToCharacter(10);
+            }
+
+            repository.deleteById(id);
+
+        }
+    }
 
     public Task findTaskById(int id) {
         for (Task task : getAllTasks()) {
@@ -59,6 +74,4 @@ public class TaskController {
             task.setCategory(newCategory);
         }
     }
-
-
 }
